move função recursiva para pasta utils
