+++ amora/cli/typer_app.py
@@ -17,27 +17,6 @@ from amora.config import settings
 from amora.dag import DependencyDAG
 from amora.models import list_models
 
-
-def recursive_dependency(
-    materialization_task,
-    model_to_task,
-):
-    """Recursively find dependencies of a materialization task."""
-
-    dependencies = materialization_task.model.__depends_on__
-
-    if not dependencies:
-        return
-    else:
-        for dependency in dependencies:
-            dependency_target_path = dependency.target_path()
-            dependency_task = materialization.Task.for_target(dependency_target_path)
-
-            model_to_task[dependency_task.model.unique_name()] = dependency_task
-
-            return recursive_dependency(dependency_task, model_to_task)
-
-
 app = typer.Typer(
     pretty_exceptions_enable=False,
     help="Amora Data Build Tool enables engineers to transform data in their warehouses "
@@ -124,7 +103,7 @@ def materialize(
         model_to_task[task.model.unique_name()] = task
 
         if depends:
-            recursive_dependency(task, model_to_task)
+            utils.recursive_dependency(task, model_to_task)
 
     dag = DependencyDAG.from_tasks(tasks=model_to_task.values())
